% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scAPAList.R
\name{calc_clusters_counts}
\alias{calc_clusters_counts}
\title{Calculates clusters' peaks counts}
\usage{
calc_clusters_counts(exp.mat, cluster.anot)
}
\arguments{
\item{exp.mat}{an scAPAList object \linkS4class{scAPAList}. Must have nonempty .cluster.anot and .cells.counts slots.}
}
\description{
Sums the reads from each peak from all the cells belonging to the same cluster.
Returns an scAPAList where the cluster counts are calculated and assigned to the clus.counts slot.
}
