% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scAPAList.R
\name{calc_cpm}
\alias{calc_cpm}
\title{Calculates the CPMs of the closter's peaks.}
\usage{
calc_cpm(x)
}
\arguments{
\item{x}{An scAPAList with a nonempty clust.count slot. 

Normalized matrix is assigned to x@norm$CPM}
}
\description{
Calculates the Counts per Milion (CPMs) of the closter's peaks.
}
