% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scAPAList.R
\name{calc_p_pui_mat}
\alias{calc_p_pui_mat}
\title{Calculates the proximal or intronic pA usage index matrix}
\usage{
calc_p_pui_mat(x, psudo = 1, int = FALSE)
}
\arguments{
\item{x}{An scAPAresults object}

\item{psudo}{numeric, the pseudo count to add before calculating PUI.
Default is 1.}

\item{in}{Weather analyzing intronic APA or not (3'UTRs). Default is FALSE.}
}
\description{
For an scAPAresults object, returns a matrix with proximal (or intronic)
pA usage index (PUI) matrix. The matrix is assigned to the slot ppui.clus.
Rownames are 3'UTRs (or genes for introns). Columns are cell clusters.
 If single cell count data is available, that is the slot ppui.cells
 is non-empty, also returns to the slot ppui.cells a matrix with the
 mean proximal/intronic PUI each cell. The average is calculated across
 all 3'UTR with more than one peak, and with reads in the given cell.
 A Cell gets NA value if it does not have such reads.
}
