% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/First_threesteps.R
\name{creat_mclus}
\alias{creat_mclus}
\title{Detects and separates bimodal peaks}
\usage{
creat_mclus(itemNumber)
}
\value{
A 'BED format' data.frame. If the peak is split, than the
data.frame contains two peaks (rows) with the corrdinates of the result
of the split. Otherwise, return the same coordinates as the original peak.
}
\description{
Adjacent pA sites may result in a single peak.
To detect and separate such peaks the R package
mclust is used to fit a Gaussian finite mixture model
with two components to the UMI counts distribution in the
interval of each peak. Mclust is used to fit an equal variance Gaussian
finite mixture model with two components to (G=2, modelNames="E") to each
list element. If the predicted means of the two fitted Gaussian components
are separated by more than three standard deviations and at least 75 nt,
the peak is split into two, according to mclust classification.
}
