% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scAPAList.R
\name{filter_IP}
\alias{filter_IP}
\title{Filter peaks that may stem from internal priming.}
\usage{
filter_IP(x, int.priming.seq, left, right)
}
\arguments{
\item{x}{An scAPAList with a nonempty down.seq slot}

\item{int.priming.seq}{Charechter, the internal priming sequence.
Defult: "AAAAAAAA".}

\item{left}{numeric, defult is 10}

\item{right}{numeric defult is 140}
}
\description{
Filter out peaks that may stem from internal priming.
If a sequance specified by int.priming.seq, is found in an interval
btween 'left' and 'right' nt
downstream the peak's 3' edge, the peak will be filtered out.
Returns a filtered scAPAList.
}
