% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/First_threesteps.R
\name{read_down.seq}
\alias{read_down.seq}
\title{Returns the genomic sequence downstream the 3' edge of a peak}
\usage{
read_down.seq(saf, char.length.path, fasta.path, chr.modify = T,
  l = -1, r = 200)
}
\arguments{
\item{saf}{A dataframe, SAF (Simplified Annotation Format) of the peaks.}

\item{char.length.path}{The path to the chromosome length file.}

\item{fasta.path}{The path to the fasta file}

\item{chr.modify}{Whether the chromosome length file has "chr"
in the chromosome names (chr1 and not 1)}

\item{l}{extract the sequence between l upstream and r
downstream the peak 3' edge. Default -1}

\item{r}{extract the sequence between l upstream and r
downstream the peak 3' edge. Default 200}
}
\value{
A data.frame with two columns: one is peak ID and the other
 is the downstream sequence.
}
\description{
This function uses bedtools getfasta to returns the 3'
edge of a peak
}
