% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scAPAList.R
\docType{class}
\name{scAPAresults-class}
\alias{scAPAresults-class}
\title{Single cell RNA-seq 3'UTR or introns analysis results data.}
\description{
A list-based S4 class for storing the results of alternative polyadenilation (APA) analysis of
single-cell RNA sequencing data.
}
\section{Slots}{

\describe{
\item{\code{.clus.counts}}{a list. Each element is a data.frame corresponding to a 3'UTR/Inton ID. Columns are cell clusters, raws are
Peaks from the 3'UTR (or intronic peak and 3'UTR counts for introns)
 Other columns (numeric) are cell barcodes, cells - read counts.}

\item{\code{.cells.counts}}{Same as .clus.counts, but columns correspond to cells.}

\item{\code{.pvalues}}{A List containing matricis. Matricis contain p-values of dynamical (APA) events.
Each matrix correspond to a comperison btwen clusters.
Rawnames are 3'UTR/Intron ID. First column is p-value, second column is FDR orreted p-value.}

\item{\code{peak.pvaues}}{A List containing matricis. Matricis contain p-values of peaks of dynamical (APA) events with multiple peaks (>3).
Each matrix correspond to a comperison btwen clusters.
Rawnames are the peak ID. First column is p-value, second column is FDR orreted p-value.}

\item{\code{.pAi.clus}}{a matrix of pA index. rawnames are UTR IDs, columns are clusters.}

\item{\code{pAi.cells}}{Same as .pAi.clus, but columns are cells.}

\item{\code{ppui.clus}}{A matrix of proximal usage index (or intronic)}
}}

