% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scAPAList.R
\name{test_APA}
\alias{test_APA}
\title{Statistical test for dynamic alternative polyadenalation events
btween cell types}
\usage{
test_APA(x, clus = "all")
}
\arguments{
\item{x}{An scAPAresults object}

\item{clus}{Charechter, the clusters to compare.
Defulte is "all" for all cluster.
To compare cluster_a abd cluster_b, set clus = c("cluster_a", "cluster_b").}
}
\description{
Uses chi-squared to test for dynamic alternative
polyadenalation events btween cell types.
Input is scAPAresult object. Output: add a matrix to the slot 'pvalues'.
 Rownames are 3'UTR IDs (or intron IDs introns). The first colomn is p-value,
  and the second is BH FDR corrected q-value.
}
