% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scAPAList.R
\name{test_peaks}
\alias{test_peaks}
\title{Tests difference in usage of peaks
between cell types}
\usage{
test_peaks(x, clus = "all", sig.level = 0.05)
}
\arguments{
\item{x}{An scAPAresults object with nonempty pvalues slot.}

\item{clus}{Character, the clusters to compare.
Default is "all" for all cluster.
To compare cluster_a to cluster_b, set clus = c("cluster_a",
"cluster_b").}

\item{sig}{test only peaks from 3'UTR with FDR q-value smaller than
this value.}
}
\description{
Uses chi-square test for goodness of fit to test for differential usage of
peaks across cell clusters. Only peaks from 3'UTR with more than 2 peaks
that came up significant. Outputs a data.frame and assigns it to the
peak.pvaues slot. The data.frame contains gene ID, 3'UTR ID peak ID,
p-value, FDR q-value, The peak usage index (PUI) for each cluster for
each peak. (from ) are given.
}
