% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/First_threesteps.R
\name{intersect_peaks}
\alias{intersect_peaks}
\title{Intersect peaks bed file with a 3'UTR or introns}
\usage{
intersect_peaks(org, path, bed.name, bedtools.path, introns = F)
}
\arguments{
\item{org}{The organism: "Hs" for human (hg19), "Mm" for mouse (mm10).}

\item{bed.name}{The path to the bed file to be intersected}

\item{bedtools.path}{The path to bedtools. (e.g "/path/to/bedtools/bin/".
End with "/", leave "" if bedtools is in your PATH environment variable)}

\item{introns}{Whether to intersect with introns or not (3'UTRs).
Default is set to FALSE.}
}
\value{
a data.frame containing the peak's GTF. Peak ID's consists of the
 gene's ensembl ID, the UTR index, and the peak index ID:
  <ensembl ID>_<UTR index>_<Peak index>. e.g: ENSMUSG00000051951.5_1_2
  has ensembl ID: ENSMUSG00000051951.5, UTR index: 1, peak index 2.
}
\description{
Intersect peaks bed file with a 3'UTR or introns. Outputs a data.frame that can be wirrten into a BED file. The function uses bedtools intersect to
intersect the bed with a 3' UTRs or introns BED. It process bedtools output into a BED format.
Lastly, it adds a peak idex to the peak ID. The most 5' peak
in the 3' UTR gets 1, and the next one 2, ect.
}
